ColorType widget working Re-implementation of  #230
