--- resources/views/components/color_picker.blade.php
@@ -1,3 +1,0 @@
-<div class="color-picker" id="{{$id}}">
-   <slider-picker v-model="color"  @changemouseup="$emit('change',$event)" />
-</div>
