--- resources/views/components/color_picker.json
@@ -1,10 +1,0 @@
-{
-    "name": "Colour Picker",
-    "data": {
-      "id": "colorPicker",
-      "color": ""
-    },
-    "include_js": true,
-    "wrapper_classes": "bg-light-grey pa4",
-    "usage_notes": "A slider element that lets a user select a color by its Hue and Luminosity values"
-}
